// src/Components/Organisms/Quotations/Drawer/types.ts
export type QuotationDrawerTypes = {
  /**Drawer type */
  type: "create" | "preview" | "edit" | "revise"
  /**Drawer visibility */
  open: boolean;
  /**Lead Id */
  leadId?: number;
  /**Quotation Id */
  quoteId?: number;
  /**Submission Id */
  submissionById?: number | null;
}